Move Page tests to ContentPage
