+++ package.json
@@ -106,6 +106,7 @@
     "koa-proxy": "^0.6.0",
     "koa-static": "^3.0.0",
     "material-ui": "^0.15.4",
+    "mobile-detect": "^1.3.3",
     "node-sass": "^3.7.0",
     "normalize.css": "^4.1.1",
     "postcss-loader": "^0.9.0",
