+++ src/components/ContentPage/index.js
@@ -1,0 +1,3 @@
+import ContentPage from './ContentPage'
+
+export default ContentPage
